+++ deploy.sh
@@ -108,25 +108,24 @@ fi
 
 # 2. Select node version
 selectNodeVersion
-echo "Running yarn config set scripts-prepend-node-path"
-eval yarn config set scripts-prepend-node-path true
+echo "Running npm config set scripts-prepend-node-path"
+eval $NPM_CMD config set scripts-prepend-node-path true
 
 # 3. Install npm packages
 if [ -e "$DEPLOYMENT_TARGET/package.json" ]; then
   cd "$DEPLOYMENT_TARGET"
-  echo "Running $NPM_CMD install --production"
-  eval yarn install --production
-  exitWithMessageOnError "yarn install failed"
+  echo "Running npm install"
+  eval $NPM_CMD install --production
+  exitWithMessageOnError "npm install failed"
   cd - > /dev/null
 fi
 
 # 4. Build the app
-
 if [ -e "$DEPLOYMENT_TARGET/package.json" ]; then
   cd "$DEPLOYMENT_TARGET"
-  echo "Building app using yarn build"
-  eval yarn build
-  exitWithMessageOnError "yarn build failed"
+  echo "Building app using npm build"
+  eval $NPM_CMD run build
+  exitWithMessageOnError "npm build failed"
   cd - > /dev/null
 fi