+++ src/__tests__/SummaryTable.test.js
@@ -16,7 +16,13 @@ const mandatoryTableRows = [
   "operatorTypeRow",
   "operatorCompanyNameRow",
   "operatorCompaniesHouseRow",
-  "operatorCharityNameRow"
+  "operatorCharityNameRow",
+  "activitiesCustomersRow",
+  "operatorContactDetailsRow",
+  "operatorEmailRow",
+  "establishmentContactDetailsRow",
+  "establishmentEmailRow",
+  "contactRepresentativeRow"
 ];
 
 // (only optional if it's optional within that page. Does not apply to pages that are optional or could be skipped.)
@@ -36,7 +42,15 @@ const testMandatoryAnswers = {
   operator_last_name: "Appleseed",
   operator_company_name: "Company name",
   operator_company_house_number: "AA123456",
-  operator_charity_name: "Charity name"
+  operator_charity_name: "Charity name",
+  customer_type: "End consumer and Other buisnesses",
+  operator_primary_number: "1234567",
+  operator_email: "operator@email.com",
+  establishment_primary_number: "12345678",
+  establishment_email: "establishment@email.com",
+  contact_representative_email: "representative@email.com",
+  contact_representative_number: "123456789",
+  contact_representative_name: "Jill"
 };
 
 // a supplementary set of all optional answer fields with example data
@@ -46,7 +60,10 @@ const testOptionalAnswers = {
   establishment_town: "Town",
   operator_street: "Street name",
   operator_town: "Town",
-  operator_charity_number: "123456"
+  contact_representative_role: "Coder",
+  operator_charity_number: "123456",
+  establishment_secondary_number: "7654321",
+  operator_secondary_number: "7654321"
 };
 
 // the complete set of possible answer fields with example data