+++ src/components/SummaryTable.js
@@ -1,6 +1,30 @@
 import React from "react";
 import { Table, Paragraph } from "govuk-react";
 import PropTypes from "prop-types";
+import * as COLOUR from "govuk-colours";
+import styled from "react-emotion";
+import {
+  FONT_SIZE,
+  LINE_HEIGHT,
+  MEDIA_QUERIES,
+  NTA_LIGHT
+} from "@govuk-react/constants";
+
+const StyledTableRow = styled("div")({
+  fontFamily: NTA_LIGHT,
+  WebkitFontSmoothing: "antialiased",
+  MozOsxFontSmoothing: "grayscale",
+  fontWeight: 400,
+  display: "inline-flex",
+  textTransform: "none",
+  fontSize: FONT_SIZE.SIZE_16,
+  lineHeight: LINE_HEIGHT.SIZE_16,
+  [MEDIA_QUERIES.LARGESCREEN]: {
+    fontSize: FONT_SIZE.SIZE_19,
+    lineHeight: LINE_HEIGHT.SIZE_19
+  },
+  color: `${COLOUR.GREY_1}`
+});
 
 const OperatorDetailsTable = props => (
   <React.Fragment>
@@ -57,6 +81,18 @@ const OperatorDetailsTable = props => (
       </Table.Row>
     ) : null}
 
+    {props.operator_first_name ? (
+      <Table.Row id="operatorNameRow">
+        <Table.CellHeader>Name</Table.CellHeader>
+        <Table.Cell>
+          <div className="bold">
+            <span id="operator_first_name">{props.operator_first_name}</span>{" "}
+            <span id="operator_last_name">{props.operator_last_name}</span>
+          </div>
+        </Table.Cell>
+      </Table.Row>
+    ) : null}
+
     {props.operator_first_line ? (
       <Table.Row id="operatorAddressRow">
         <Table.CellHeader>Operator address</Table.CellHeader>
@@ -71,13 +107,17 @@ const OperatorDetailsTable = props => (
       </Table.Row>
     ) : null}
 
-    {props.operator_first_name ? (
-      <Table.Row id="operatorNameRow">
-        <Table.CellHeader>Name</Table.CellHeader>
+    {props.operator_primary_number ? (
+      <Table.Row id="operatorContactDetailsRow">
+        <Table.CellHeader>Contact number</Table.CellHeader>
         <Table.Cell>
           <div className="bold">
-            <span id="operator_first_name">{props.operator_first_name}</span>{" "}
-            <span id="operator_last_name">{props.operator_last_name}</span>
+            <div id="operator_primary_number">
+              {props.operator_primary_number}
+            </div>
+            <div id="operator_secondary_number">
+              {props.operator_secondary_number || null}
+            </div>
           </div>
         </Table.Cell>
       </Table.Row>
@@ -89,6 +129,43 @@ const OperatorDetailsTable = props => (
         <Table.Cell>
           <div className="bold" id="operator_type">
             {props.operator_type}
+          </div>
+        </Table.Cell>
+      </Table.Row>
+    ) : null}
+
+    {props.operator_email ? (
+      <Table.Row id="operatorEmailRow">
+        <Table.CellHeader>Email address</Table.CellHeader>
+        <Table.Cell>
+          <div className="bold" id="operator_email">
+            {props.operator_email}
+          </div>
+        </Table.Cell>
+      </Table.Row>
+    ) : null}
+
+    {props.contact_representative_email ? (
+      <Table.Row id="contactRepresentativeRow">
+        <Table.CellHeader>Designated contact</Table.CellHeader>
+        <Table.Cell>
+          <StyledTableRow>
+            <div display id="contact_representative_name">
+              {props.contact_representative_name}
+            </div>
+          </StyledTableRow>
+          <StyledTableRow>
+            <div id="contact_representative_role">
+              {props.contact_representative_role
+                ? `\u2000\u2022 ${props.contact_representative_role}`
+                : null}
+            </div>
+          </StyledTableRow>
+          <div className="bold" id="contact_representative_number">
+            {props.contact_representative_number}
+          </div>
+          <div className="bold" id="contact_representative_email">
+            {props.contact_representative_email}
           </div>
         </Table.Cell>
       </Table.Row>
@@ -139,6 +216,33 @@ const EstablishmentDetailsTable = props => (
         </Table.Cell>
       </Table.Row>
     ) : null}
+
+    {props.establishment_primary_number ? (
+      <Table.Row id="establishmentContactDetailsRow">
+        <Table.CellHeader>Contact number</Table.CellHeader>
+        <Table.Cell>
+          <div className="bold">
+            <div id="establishment_primary_number">
+              {props.establishment_primary_number}
+            </div>
+            <div id="establishment_secondary_number">
+              {props.establishment_secondary_number || null}
+            </div>
+          </div>
+        </Table.Cell>
+      </Table.Row>
+    ) : null}
+
+    {props.establishment_email ? (
+      <Table.Row id="establishmentEmailRow">
+        <Table.CellHeader>Email address</Table.CellHeader>
+        <Table.Cell>
+          <div className="bold" id="establishment_email">
+            {props.establishment_email}
+          </div>
+        </Table.Cell>
+      </Table.Row>
+    ) : null}
   </React.Fragment>
 );
 
@@ -152,6 +256,17 @@ const FoodActivitiesTable = props => (
         <div />
       </Table.Cell>
     </Table.Row>
+
+    {props.customer_type ? (
+      <Table.Row id="activitiesCustomersRow">
+        <Table.CellHeader>Customers</Table.CellHeader>
+        <Table.Cell>
+          <div className="bold" id="customer_type">
+            {props.customer_type}
+          </div>
+        </Table.Cell>
+      </Table.Row>
+    ) : null}
   </React.Fragment>
 );
 
@@ -185,5 +300,6 @@ SummaryTable.propTypes = {
   establishment_first_line: PropTypes.string,
   establishment_street: PropTypes.string,
   establishment_town: PropTypes.string,
-  establishment_postcode: PropTypes.string
+  establishment_postcode: PropTypes.string,
+  customer_type: PropTypes.string
 };