+++ src/server/controllers/back.controller.js
@@ -1,16 +1,12 @@
-const pathJSON = require("../services/path.json");
 const { moveAlongPath, editPath } = require("../services/path.service");
 
-const backController = (currentPage, previousAnswers) => {
+const backController = (currentPage, previousAnswers = {}) => {
   let previousPage;
+  let newPath;
 
-  if (previousAnswers) {
-    const answersArray = Object.values(previousAnswers);
-    const newPath = editPath(pathJSON, answersArray, currentPage);
-    previousPage = moveAlongPath(newPath, currentPage, -1);
-  } else {
-    previousPage = moveAlongPath(pathJSON, currentPage, -1);
-  }
+  const answersArray = Object.values(previousAnswers);
+  newPath = editPath(answersArray, currentPage);
+  previousPage = moveAlongPath(newPath, currentPage, -1);
 
   return previousPage;
 };