+++ src/server/controllers/continue.controller.js
@@ -1,38 +1,44 @@
-const pathJSON = require("../services/path.json");
 const { moveAlongPath, editPath } = require("../services/path.service");
 const { validate } = require("../services/validation.service");
 const {
-  transformAnswersForSubmit
-} = require("../services/data-transform.service");
-const {
   cleanInactivePathAnswers,
-  cleanEmptiedAnswers
+  cleanEmptiedAnswers,
+  cleanSwitches
 } = require("../services/session-management.service");
 
-const continueController = (currentPage, previousAnswers, newAnswers) => {
+const continueController = (
+  currentPage,
+  previousAnswers,
+  newAnswers = {},
+  switches
+) => {
   const controllerResponse = {
     validatorErrors: {},
     redirectRoute: null,
-    cumulativeAnswers: {}
+    cumulativeAnswers: {},
+    switches: {}
   };
 
   const newAnswersArray = Object.values(newAnswers);
 
   let cleanedPreviousAnswers = Object.assign({}, previousAnswers);
 
-  if (newAnswersArray.length > 0) {
-    // remove any answers that were previously given a truthy value but have since been emptied
-    cleanedPreviousAnswers = cleanEmptiedAnswers(
-      previousAnswers,
-      newAnswersArray,
-      currentPage
-    );
-  }
+  // remove any answers that were previously given a truthy value but have since been emptied
+  cleanedPreviousAnswers = cleanEmptiedAnswers(
+    previousAnswers,
+    newAnswersArray,
+    currentPage
+  );
 
   controllerResponse.cumulativeAnswers = Object.assign(
     {},
     cleanedPreviousAnswers,
     newAnswers
+  );
+
+  controllerResponse.switches = Object.assign(
+    {},
+    cleanSwitches(controllerResponse.cumulativeAnswers, switches)
   );
 
   controllerResponse.validatorErrors = Object.assign(
@@ -50,7 +56,7 @@ const continueController = (currentPage, previousAnswers, newAnswers) => {
     controllerResponse.cumulativeAnswers
   );
 
-  const newPath = editPath(pathJSON, cumulativeAnswersArray, currentPage);
+  const newPath = editPath(cumulativeAnswersArray, currentPage);
 
   // remove any answers that are associated with an inactive page on the path
   controllerResponse.cumulativeAnswers = cleanInactivePathAnswers(